requirements.txt
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
joblib==1.4.2
influxdb-client==1.44.0
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
prometheus-client==0.20.0
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.4.2
seaborn==0.13.2
pyzmq==26.0.3
